+++ Dungeons of Daggorath/viewer.h
@@ -166,7 +166,7 @@ private:
 	// Internal Implementation
 	void drawVectorListAQ(int VLA[]);
 	void drawCharacter(char c);
-	void drawString(int x, int y, const char * str, int len);
+	void drawString(int x, int y, const char * str, size_t len);
 	void drawString_internal(int x, int y, dodBYTE * str, int len);
 	void plotPoint(double X, double Y);
 	char dod_to_ascii(dodBYTE c);
